Move common list styling to List component SASS
