+++ src/components/lists/CallList.scss
@@ -1,10 +1,4 @@
 .CallList {
-
-    width: 100%;
-
-    @include medium-screen {
-        max-width: 540px;
-    }
     .ListHeader {
         li:nth-child(1) {
             @include col(7,12);
