+++ src/components/lists/List.scss
@@ -1,0 +1,19 @@
+.List {
+    width: 100%;
+
+    .List-item {
+        width: 100%;
+        margin: 1em 0;
+        padding: 1em;
+        @include card;
+
+        &:hover {
+            background-color: darken($c-ui-bg, 1);
+            cursor: pointer;
+        }
+    }
+
+    @include medium-screen {
+        max-width: 540px;
+    }
+}
