+++ src/components/lists/items/CallListItem.scss
@@ -1,14 +1,4 @@
 .CallListItem {
-    width: 100%;
-    margin: 1em 0;
-    padding: 1em;
-    @include card;
-
-    &:hover {
-        background-color: darken($c-ui-bg, 1);
-        cursor: pointer;
-    }
-
     .Avatar {
         @include col(1.5,12, $align:middle);
         &:first-child {
